필요 없는 파일 삭제

+++ src/main/java/kr/co/itwillbs/de/mes/dto/RawMaterialSearchDTO.java
@@ -1,7 +1,5 @@
 package kr.co.itwillbs.de.mes.dto;
 
-import java.math.BigDecimal;
-
 import kr.co.itwillbs.de.common.vo.PageDTO;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -26,8 +24,8 @@ public class RawMaterialSearchDTO {
     private Integer minQuantity;    // 최소 수량
     private Integer maxQuantity;    // 최대 수량
     
-    private BigDecimal minPrice;    // 최소 가격
-    private BigDecimal maxPrice;    // 최대 가격
+    private Integer minPrice;    // 최소 가격
+    private Integer maxPrice;    // 최대 가격
     
     private PageDTO pageDTO = new PageDTO();  // 페이징 정보
 }
